+++ pkzo/Mesh.cpp
@@ -29,11 +29,6 @@
 
 namespace pkzo
 {
-    std::shared_ptr<Mesh> Mesh::load_file(const std::filesystem::path& file)
-    {
-        return std::make_shared<Mesh>();
-    }
-
     std::shared_ptr<Mesh> Mesh::create_plane(const glm::vec2& size)
     {
         const auto hs = size * 0.5f;
@@ -51,19 +46,14 @@ namespace pkzo
         return mesh;
     }
 
-    Mesh::Mesh(const std::filesystem::path& file)
-    {
-
-    }
-
     Mesh::~Mesh()
     {
-        free_videomemory();
+        release();
     }
 
     void Mesh::add_vertex(const glm::vec3& vertex, const glm::vec3& normal, const glm::vec3& tangent, const glm::vec2& texcoord)
     {
-        vertices.push_back(glm::vec4(vertex, 1.0f));
+        vertexes.push_back(vertex);
         normals.push_back(normal);
         tangents.push_back(tangent);
         texcoords.push_back(texcoord);
@@ -71,7 +61,17 @@ namespace pkzo
 
     void Mesh::add_triangle(uint a, uint b, uint c)
     {
-        indexes.push_back({a, b, c});
+        triangles.push_back({a, b, c});
+    }
+
+    unsigned int Mesh::get_vertex_count() const noexcept
+    {
+        return vertexes.size();
+    }
+
+    unsigned int Mesh::get_triangle_count() const noexcept
+    {
+        return triangles.size();
     }
 
     void Mesh::upload()
@@ -83,13 +83,13 @@ namespace pkzo
 
         glGenVertexArrays(1, &vao);
         glCreateBuffers(static_cast<GLsizei>(buffers.size()), buffers.data());
-        glNamedBufferData(buffers[to_underlying(BufferId::VERTEX)], vertices.size() * sizeof(glm::vec3), glm::value_ptr(vertices.front()), GL_STATIC_DRAW);
+        glNamedBufferData(buffers[to_underlying(BufferId::VERTEX)], vertexes.size() * sizeof(glm::vec3), glm::value_ptr(vertexes.front()), GL_STATIC_DRAW);
         glNamedBufferData(buffers[to_underlying(BufferId::NORMAL)], normals.size() * sizeof(glm::vec3), glm::value_ptr(normals[0]), GL_STATIC_DRAW);
         glNamedBufferData(buffers[to_underlying(BufferId::TANGENT)], tangents.size() * sizeof(glm::vec3), glm::value_ptr(tangents[0]), GL_STATIC_DRAW);
         glNamedBufferData(buffers[to_underlying(BufferId::TEXCOORD)], texcoords.size() * sizeof(glm::vec2), glm::value_ptr(texcoords[0]), GL_STATIC_DRAW);
 
         glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, buffers[to_underlying(BufferId::INDEX)]);
-        glBufferData(GL_ELEMENT_ARRAY_BUFFER, indexes.size() * 3 * sizeof(unsigned int), glm::value_ptr(indexes[0]), GL_STATIC_DRAW);
+        glBufferData(GL_ELEMENT_ARRAY_BUFFER, triangles.size() * 3 * sizeof(unsigned int), glm::value_ptr(triangles[0]), GL_STATIC_DRAW);
 
         check_glerror();
     }
@@ -99,7 +99,7 @@ namespace pkzo
         return vao != 0;
     }
 
-    void Mesh::free_videomemory()
+    void Mesh::release()
     {
         if (! is_uploaded())
         {
@@ -153,7 +153,7 @@ namespace pkzo
     void Mesh::draw()
     {
         glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, buffers[to_underlying(BufferId::INDEX)]);
-        glDrawElements(GL_TRIANGLES, static_cast<GLsizei>(indexes.size() * 3u), GL_UNSIGNED_INT, 0);
+        glDrawElements(GL_TRIANGLES, static_cast<GLsizei>(triangles.size() * 3u), GL_UNSIGNED_INT, 0);
         glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
     }